+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# development version
+
 # callr 3.4.1
 
 * callr is now more careful when loading the local `.Rprofile` in the